pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fibz-bot"
version = "0.4.0"
description = "Fibz — Python Discord bot with Vertex AI (Gemini), ChromaDB, consent policies, tools, OCR, GCS, and PDF/image citation."
authors = [{name = "You"}]
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.1",
    "google-cloud-aiplatform>=1.70.0",
    "vertexai>=1.70.0",
    "chromadb>=0.5.5",
    "numpy>=1.26.0",
    "pillow>=10.2.0",
    "pypdf>=4.2.0",
    "python-docx>=1.1.0",
    "python-pptx>=0.6.23",
    "ffmpeg-python>=0.2.0",
    "pyyaml>=6.0.1",
    "google-cloud-speech>=2.27.0",
    "google-cloud-storage>=2.17.0",
    "google-cloud-vision>=3.7.3",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.5.0",
    "black>=24.3.0",
    "mypy>=1.10.0",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP"]
ignore = []
fix = true

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
